+++ Jint/Collections/ListDictionary.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Runtime.CompilerServices;
 using Jint.Runtime;
 
 namespace Jint.Collections
