+++ Jint/Native/Map/MapConstructor.cs
@@ -41,7 +41,7 @@ namespace Jint.Native.Map
 
         protected override void Initialize()
         {
-            var symbols = new PropertyDictionary(1)
+            var symbols = new SymbolDictionary(1)
             {
                 [GlobalSymbolRegistry.Species] = new GetSetPropertyDescriptor(get: new ClrFunctionInstance(_engine, "get [Symbol.species]", Species, 0, PropertyFlag.Configurable), set: Undefined, PropertyFlag.Configurable)
             };
