+++ Jint/Native/Set/SetConstructor.cs
@@ -41,7 +41,7 @@ namespace Jint.Native.Set
 
         protected override void Initialize()
         {
-            var symbols = new PropertyDictionary(1)
+            var symbols = new SymbolDictionary(1)
             {
                 [GlobalSymbolRegistry.Species] = new GetSetPropertyDescriptor(get: new ClrFunctionInstance(_engine, "get [Symbol.species]", Species, 0, PropertyFlag.Configurable), set: Undefined, PropertyFlag.Configurable)
             };
